testing

+++ test_neon2.c
@@ -89,8 +89,8 @@ void main(){
      ///     PIVOT    ///
     ////////////////////
     
-    int8_t first_op;
-    int8_t second_op;    
+    int16_t first_op;
+    int16_t second_op;    
 
     for(k = 0; k < n; k++){
         first_op = A[ k*m + k ];
